refactor app into modules.
