+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "axios": "^0.17.0",
+    "eslint-config-react-app": "^2.1.0",
     "google-map-react": "^0.25.0",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
