+++ src/components/App/App.js
@@ -2,16 +2,18 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { createStore, applyMiddleware } from 'redux';
 import promise from 'redux-promise';
-import reducers from '../reducers';
+import reducers from '../../reducers';
 
 import './App.css';
-import MapContainer from '../containers/MapContainer';
+import {
+  Map,
+} from './components';
 
 const createStoreWithMiddleware = applyMiddleware(promise)(createStore);
 
 const App = () => (
   <Provider store={createStoreWithMiddleware(reducers)}>
-    <MapContainer />
+    <Map />
   </Provider>
 );
