+++ src/components/App/components/Map/Map.js
@@ -1,7 +1,9 @@
 import React, { Component } from 'react';
 import GoogleMap from 'google-map-react';
 
-import SpaceStation from './SpaceStation';
+import {
+  SpaceStation,
+} from './components';
 import shouldPureComponentUpdate from 'react-pure-render/function';
 import './Map.css';
 
@@ -16,7 +18,7 @@ class Map extends Component {
   onChildClick = (key, childProps) => {
     this.props.toggleLock();
   };
-
+  _onClick = ({x, y, lat, lng, event}) => console.log(x, y, lat, lng, event)
   shouldComponentUpdate = shouldPureComponentUpdate;
 
   render() {
@@ -27,6 +29,7 @@ class Map extends Component {
           center={this.props.iss.center}
           zoom={8}
           onChildClick={this.onChildClick}
+          onClick={this._onClick}
         >
           <SpaceStation {...this.props.iss.coords} />
         </GoogleMap>
