+++ src/components/App/components/Map/components/SpaceStation/SpaceStation.css
