+++ src/components/App/components/Map/components/SpaceStation/SpaceStation.js
@@ -5,7 +5,7 @@ import './SpaceStation.css';
 const SpaceStation = () => (
   <div className="iss">
     <p className="title">Click the ISS toggle lock on</p>
-    <img src={require('../images/iss.png')} alt="iss" />
+    <img src={require('../../../../../../images/iss.png')} alt="iss" />
   </div>
 );
