+++ src/components/App/components/Map/components/SpaceStation/index.js
@@ -1,0 +1,3 @@
+import SpaceStation from './SpaceStation'
+
+export default SpaceStation;
