+++ src/components/App/components/Map/components/index.js
@@ -1,0 +1,1 @@
+export { default as SpaceStation } from './SpaceStation';
