+++ src/components/App/components/Map/index.js
@@ -1,0 +1,3 @@
+import MapContainer from './MapContainer'
+
+export default MapContainer;
