+++ src/components/App/components/index.js
@@ -1,0 +1,1 @@
+export { default as Map } from './Map';
